Walshy fixes
Moved modlog events to their own class
Changed from a list to a set
