+++ src/main/java/stream/flarebot/flarebot/Events.java
@@ -120,11 +120,6 @@ public class Events extends ListenerAdapter {
 
     @Override
     public void onGuildMemberJoin(GuildMemberJoinEvent event) {
-        FlareBotManager.getInstance().getGuild(event.getGuild().getId())
-                .getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("User Join")
-                .addField("User", MessageUtils.getTag(event.getUser()) + " (" + event.getUser().getId() + ")", true)
-                .build(), ModlogEvent.MEMBER_JOIN);
         PlayerCache cache = flareBot.getPlayerCache(event.getMember().getUser().getId());
         cache.setLastSeen(LocalDateTime.now());
         GuildWrapper wrapper = FlareBotManager.getInstance().getGuild(event.getGuild().getId());
@@ -177,15 +172,6 @@ public class Events extends ListenerAdapter {
         }
     }
 
-    @Override
-    public void onGuildMemberLeave(GuildMemberLeaveEvent event) {
-        FlareBotManager.getInstance().getGuild(event.getGuild().getId())
-                .getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("User Leave")
-                .addField("User", MessageUtils.getTag(event.getUser()) + " (" + event.getUser().getId() + ")", true)
-                .build(), ModlogEvent.MEMBER_LEAVE);
-    }
-
     private void handle(Throwable e1, GuildMemberJoinEvent event, List<Role> roles) {
         if (!e1.getMessage().startsWith("Can't modify a role with higher")) {
             MessageUtils.sendPM(event.getGuild().getOwner().getUser(),
@@ -233,12 +219,6 @@ public class Events extends ListenerAdapter {
 
     @Override
     public void onGuildVoiceJoin(GuildVoiceJoinEvent event) {
-        FlareBotManager.getInstance().getGuild(event.getGuild().getId())
-                .getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("Voice join")
-                .addField("User", MessageUtils.getTag(event.getMember().getUser()) + " (" + event.getMember().getUser().getId() + ")", true)
-                .addField("Channel", event.getChannelJoined().getName() + " (" + event.getChannelJoined().getId() + ")", true)
-                .build(), ModlogEvent.MEMBER_VOICE_JOIN);
         if (event.getMember().getUser().equals(event.getJDA().getSelfUser())) {
             if (FlareBot.getInstance().getMusicManager().hasPlayer(event.getGuild().getId())) {
                 FlareBot.getInstance().getMusicManager().getPlayer(event.getGuild().getId()).setPaused(false);
@@ -248,12 +228,6 @@ public class Events extends ListenerAdapter {
 
     @Override
     public void onGuildVoiceLeave(GuildVoiceLeaveEvent event) {
-        FlareBotManager.getInstance().getGuild(event.getGuild().getId())
-                .getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("Voice leave")
-                .addField("User", MessageUtils.getTag(event.getMember().getUser()) + " (" + event.getMember().getUser().getId() + ")", true)
-                .addField("Channel", event.getChannelLeft().getName() + " (" + event.getChannelLeft().getId() + ")", true)
-                .build(), ModlogEvent.MEMBER_VOICE_LEAVE);
         if (event.getMember().getUser().getIdLong() == event.getJDA().getSelfUser().getIdLong()) {
             if (FlareBot.getInstance().getMusicManager().hasPlayer(event.getGuild().getId())) {
                 FlareBot.getInstance().getMusicManager().getPlayer(event.getGuild().getId()).setPaused(true);
@@ -351,111 +325,9 @@ public class Events extends ListenerAdapter {
     }
 
     @Override
-    public void onRoleCreate(RoleCreateEvent event) {
-        event.getGuild().getAuditLogs().queue(auditLog -> {
-            AuditLogEntry entry = auditLog.get(0);
-            FlareBotManager.getInstance().getGuild(event.getGuild().getId())
-                    .getAutoModConfig().postToModLog(new EmbedBuilder()
-                    .setTitle("Role create")
-                    .addField("Role", event.getRole().getName() + " (" + event.getRole().getId() + ")", true)
-                    .addField("Responsible moderator", entry.getUser().getAsMention(), true)
-                    .build(), ModlogEvent.ROLE_CREATE);
-        });
-
-    }
-
-    @Override
     public void onRoleDelete(RoleDeleteEvent event) {
-        event.getGuild().getAuditLogs().queue(auditLog -> {
-            AuditLogEntry entry = auditLog.get(0);
-            FlareBotManager.getInstance().getGuild(event.getGuild().getId())
-                    .getAutoModConfig().postToModLog(new EmbedBuilder()
-                    .setTitle("Role delete")
-                    .addField("Role", event.getRole().getName() + " (" + event.getRole().getId() + ")", true)
-                    .addField("Responsible moderator", entry.getUser().getAsMention(), true)
-                    .build(), ModlogEvent.ROLE_DELETE);
-        });
         if (FlareBotManager.getInstance().getGuild(event.getGuild().getId()).getSelfAssignRoles().contains(event.getRole().getId())) {
             FlareBotManager.getInstance().getGuild(event.getGuild().getId()).getSelfAssignRoles().remove(event.getRole().getId());
-        }
-    }
-    private long responseNumber = 0;
-    @Override
-    public void onGenericRoleUpdate(GenericRoleUpdateEvent event) {
-        if (event instanceof RoleUpdatePositionEvent) {
-            int oldpos = ((RoleUpdatePositionEvent) event).getOldPosition();
-            int newpos = event.getRole().getPosition();
-            if (oldpos == newpos)
-                return;
-            EmbedBuilder embedBuilder = new EmbedBuilder();
-            embedBuilder.setTitle("Role Move");
-            embedBuilder.addField("Role", event.getRole().getName() + " (" + event.getRole().getId() + ")", true);
-            embedBuilder.addField("Position", "`" + oldpos + "` -> `" + newpos + "`", true);
-            List<Role> roles = event.getGuild().getRoles();
-            String surounding = "```md\n";
-            if (newpos != roles.size())
-                surounding += "+ " + roles.get(roles.size() - (newpos + 3)).getName() + "\n";
-           surounding += "> " + event.getRole().getName() + "\n";
-            if (newpos != 0) {
-                surounding += "+ " + roles.get(roles.size() - (newpos + 1)).getName() + "\n";
-            }
-            surounding += "```";
-            embedBuilder.addField("Surrounding roles", surounding, false);
-            FlareBotManager.getInstance().getGuild(event.getGuild().getId()).getAutoModConfig().postToModLog(embedBuilder.build(), ModlogEvent.ROLE_MOVE);
-        } else {
-            if (event.getResponseNumber() == responseNumber) {
-                return;
-            }
-            responseNumber = event.getResponseNumber();
-            event.getGuild().getAuditLogs().limit(1).queue(auditLogs -> {
-                AuditLogEntry entry = auditLogs.get(0);
-                Map<String, AuditLogChange> changes = entry.getChanges();
-                EmbedBuilder permissionsBuilder = new EmbedBuilder();
-                permissionsBuilder.setTitle("Role Change");
-
-                permissionsBuilder.addField("Role", event.getRole().getName() + " (" + event.getRole().getId() + ")", false);
-
-                System.out.println(changes.keySet());
-
-                if (changes.containsKey("permissions")) {
-                    AuditLogChange change = changes.get("permissions");
-                    Map<Boolean, List<Permission>> permChanges = GeneralUtils.getChangedPerms(Permission.getPermissions(((Integer) change.getOldValue()).longValue()), Permission.getPermissions(((Integer) change.getNewValue()).longValue()));
-                    if (permChanges.get(true).size() > 0) {
-                        StringBuilder added = new StringBuilder();
-                        for (Permission addedPerm : permChanges.get(true)) {
-                            added.append(addedPerm.getName()).append("\n");
-                        }
-                        permissionsBuilder.addField("Added Perms", "```\n" + added.toString() + "```", false);
-                    }
-                    if (permChanges.get(false).size() > 0) {
-                        StringBuilder removed = new StringBuilder();
-                        for (Permission removedPerm : permChanges.get(false)) {
-                            removed.append(removedPerm.getName()).append("\n");
-                        }
-                        permissionsBuilder.addField("Removed Perms", "```\n" + removed.toString() + "```", false);
-                    }
-                }
-                if (changes.containsKey("name")) {
-                    AuditLogChange change = changes.get("name");
-                    permissionsBuilder.addField("Name Change", "`" + change.getOldValue() + "` -> `" + change.getNewValue() + "`", true);
-                }
-                if (changes.containsKey("mentionable")) {
-                    AuditLogChange change = changes.get("mentionable");
-                    permissionsBuilder.addField("Mentionable", "`" + change.getNewValue() + "`", true);
-                }
-                if (changes.containsKey("hoist")) {
-                    AuditLogChange change = changes.get("hoist");
-                    permissionsBuilder.addField("Displayed Separately", "`" + change.getNewValue() + "`", true);
-                }
-                if (changes.containsKey("color")) {
-                    AuditLogChange change = changes.get("color");
-                    permissionsBuilder.addField("Color Change", "`#" + Integer.toHexString(change.getOldValue()) + "` -> `#" + Integer.toHexString(change.getNewValue()) + "`", true);
-
-                }
-                permissionsBuilder.addField("Responsible Moderator", entry.getUser().getAsMention(), true);
-                FlareBotManager.getInstance().getGuild(event.getGuild().getId()).getAutoModConfig().postToModLog(permissionsBuilder.build(), ModlogEvent.ROLE_EDIT);
-
-            });
         }
     }
 
@@ -464,85 +336,6 @@ public class Events extends ListenerAdapter {
         Guild guild = event.getGuild();
         AuditLogEntry entry = guild.getAuditLogs().complete().get(0);
         FlareBotManager.getInstance().getGuild(guild.getId()).getAutoModConfig().postToModLog(event.getUser(), entry.getUser(), new Punishment(ModlogAction.BAN), true);
-    }
-
-    @Override
-    public  void onGuildMemberRoleAdd(GuildMemberRoleAddEvent event) {
-        AuditLogEntry entry = event.getGuild().getAuditLogs().complete().get(0);
-        Map<String, AuditLogChange> changes = entry.getChanges();
-        AuditLogChange change = changes.get("$add");
-        @SuppressWarnings("unchecked")
-        HashMap<String, String> role = ((ArrayList<HashMap<String, String>>)change.getNewValue()).get(0);
-
-        FlareBotManager.getInstance().getGuild(entry.getGuild().getId())
-                .getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("Role add")
-                .addField("User", MessageUtils.getTag(event.getUser()) + " (" + event.getUser().getId() + ")", true)
-                .addField("Role", role.get("name") + " (" + role.get("id") + ")", true)
-                .addField("Responsible moderator", entry.getUser().getAsMention(), true)
-                .build(), ModlogEvent.MEMBER_ROLE_GIVE);
-    }
-
-    @Override
-    public void onGuildMemberRoleRemove(GuildMemberRoleRemoveEvent event) {
-        AuditLogEntry entry = event.getGuild().getAuditLogs().complete().get(0);
-        Map<String, AuditLogChange> changes = entry.getChanges();
-        AuditLogChange change = changes.get("$remove");
-        @SuppressWarnings("unchecked")
-        HashMap<String, String> role = ((ArrayList<HashMap<String, String>>)change.getNewValue()).get(0);
-
-        FlareBotManager.getInstance().getGuild(entry.getGuild().getId())
-                .getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("Role remove")
-                .addField("User", MessageUtils.getTag(event.getUser()) + " (" + event.getUser().getId() + ")", true)
-                .addField("Role", role.get("name") + " (" + role.get("id") + ")", true)
-                .addField("Responsible moderator", entry.getUser().getAsMention(), true)
-                .build(), ModlogEvent.MEMBER_ROLE_REMOVE);
-    }
-
-    @Override
-    public void onTextChannelCreate(TextChannelCreateEvent event) {
-        handleChannelCreate(FlareBotManager.getInstance().getGuild(event.getGuild().getId()), event.getChannel());
-    }
-
-    @Override
-    public  void onVoiceChannelCreate(VoiceChannelCreateEvent event) {
-        handleChannelCreate(FlareBotManager.getInstance().getGuild(event.getGuild().getId()), event.getChannel());
-    }
-
-    @Override
-    public void onTextChannelDelete(TextChannelDeleteEvent event) {
-        handleChannelDelete(FlareBotManager.getInstance().getGuild(event.getGuild().getId()), event.getChannel());
-    }
-
-    @Override
-    public void onVoiceChannelDelete(VoiceChannelDeleteEvent event) {
-        handleChannelDelete(FlareBotManager.getInstance().getGuild(event.getGuild().getId()), event.getChannel());
-    }
-
-    @Override
-    public void onMessageUpdate(MessageUpdateEvent event) {
-        //TODO message caching
-    }
-
-    private void handleChannelCreate(GuildWrapper wrapper, Channel channel) {
-        AuditLogEntry entry = wrapper.getGuild().getAuditLogs().complete().get(0);
-        wrapper.getAutoModConfig().postToModLog(new EmbedBuilder()
-        .setTitle("Channel create")
-        .addField("Type", channel.getType().name().toLowerCase(), true)
-        .addField("Name", channel.getName(), true)
-        .addField("Responsible moderator", entry.getUser().getAsMention(), true)
-        .build(), ModlogEvent.CHANNEL_CREATE);
-    }
-
-    private void handleChannelDelete(GuildWrapper wrapper, Channel channel) {
-        AuditLogEntry entry = wrapper.getGuild().getAuditLogs().complete().get(0);
-        wrapper.getAutoModConfig().postToModLog(new EmbedBuilder()
-                .setTitle("Channel delete")
-                .addField("Type", channel.getType().name().toLowerCase(), true)
-                .addField("Name", channel.getName(), true)
-                .addField("Responsible moderator", entry.getUser().getAsMention(), true)
-                .build(), ModlogEvent.CHANNEL_DELETE);
     }
 
     private void handleCommand(GuildMessageReceivedEvent event, Command cmd, String[] args) {
